絵文字入力エリアをコンポーネント化
