+++ frontend/src/components/main/globalParts/UserIcon.jsx
@@ -8,7 +8,6 @@ const UserIcon=({image,color,size})=>{
     const fontAwesomeValue=iconsCatalog.image[image].fontAwesomeValue;
     const tailwindClass=iconsCatalog.color[color].tailwindClass;
     
-
     return (
         <div className={size?"w-"+size+" h-"+size+" rounded-full m-auto  border border-gray-200 bg-white  flex items-center justify-center":""}>
             {fontAwesomeValue?<FontAwesomeIcon icon={fontAwesomeValue+" fa-solid"} className={"rounded-full w-2/3 h-2/3 "+tailwindClass} />:""}
